v 0.132 analyze_candles --> ef-AND-rsi, trend_near не равен нулю analyze_events --> убрал break в цикле определения long-short

+++ stop_tbot.py
@@ -6,6 +6,7 @@ import datetime
 import write2file
 
 pid_to_kill = sql2data.pid_from_sql('stream_connection')
-os.kill(pid_to_kill, signal.SIGINT)
-data2sql.update_pid('stream_connection', 0)
-write2file.write(str(datetime.datetime.now())[:19] + ' STOP stream_connection', 'log.txt')
+if pid_to_kill:
+    os.kill(pid_to_kill, signal.SIGINT)
+    data2sql.update_pid('stream_connection', 0)
+    write2file.write(str(datetime.datetime.now())[:19] + ' STOP stream_connection', 'log.txt')
